requirements.txt
Flask==2.3.2
langchain==0.0.235
openai==0.27.8
pdfminer.six==20221105
PyPDF2==3.0.1
python-dotenv==1.0.0
